refactor: rename post_stream to eventsource and simplify URL/header handling
